+++ CHANGES.md
@@ -1,5 +1,10 @@
 ## Changelog
 
+### [1.0.1] - 2016-06-03
+#### Fixed
+- iOS7 crash on start
+- iOS attempt to fix #46 and #39
+
 ### [1.0.0] - 2016-06-01
 #### Added
 - Android ANDROID_FUSED_LOCATION stopOnStillActivity (enhancement #69)